# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinAegis Core Banking API",
        "description": "Open Source Core Banking as a Service - A modern, scalable, and secure core banking platform built with Laravel 12, featuring event sourcing, domain-driven design, workflow orchestration, and quantum-resistant security measures.",
        "contact": {
            "name": "FinAegis Support",
            "email": "support@finaegis.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/accounts/{uuid}/balances": {
            "get": {
                "tags": [
                    "Account Balances"
                ],
                "summary": "Get account balances",
                "description": "Retrieve all asset balances for a specific account",
                "operationId": "getAccountBalances",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The account UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    },
                    {
                        "name": "asset",
                        "in": "query",
                        "description": "Filter by specific asset code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "positive",
                        "in": "query",
                        "description": "Only show positive balances",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "account_uuid": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "balances": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "asset_code": {
                                                                "type": "string"
                                                            },
                                                            "balance": {
                                                                "type": "integer"
                                                            },
                                                            "formatted": {
                                                                "type": "string"
                                                            },
                                                            "asset": {
                                                                "properties": {
                                                                    "code": {
                                                                        "type": "string"
                                                                    },
                                                                    "name": {
                                                                        "type": "string"
                                                                    },
                                                                    "type": {
                                                                        "type": "string"
                                                                    },
                                                                    "symbol": {
                                                                        "type": "string"
                                                                    },
                                                                    "precision": {
                                                                        "type": "integer"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "summary": {
                                                    "properties": {
                                                        "total_assets": {
                                                            "type": "integer"
                                                        },
                                                        "total_usd_equivalent": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account not found"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "The specified account UUID was not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/balances": {
            "get": {
                "tags": [
                    "Account Balances"
                ],
                "summary": "List all account balances",
                "description": "Get balances across all accounts with filtering and aggregation options",
                "operationId": "listAllBalances",
                "parameters": [
                    {
                        "name": "asset",
                        "in": "query",
                        "description": "Filter by specific asset code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "min_balance",
                        "in": "query",
                        "description": "Minimum balance filter (in smallest unit)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1000
                        }
                    },
                    {
                        "name": "user_uuid",
                        "in": "query",
                        "description": "Filter by account owner",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of results per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "account_uuid": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "asset_code": {
                                                        "type": "string"
                                                    },
                                                    "balance": {
                                                        "type": "integer"
                                                    },
                                                    "formatted": {
                                                        "type": "string"
                                                    },
                                                    "account": {
                                                        "properties": {
                                                            "uuid": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "user_uuid": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total_accounts": {
                                                    "type": "integer"
                                                },
                                                "total_balances": {
                                                    "type": "integer"
                                                },
                                                "asset_totals": {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string",
                                                        "property": null
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "title": "Validation Error Response",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ],
                            "amount": [
                                "The amount must be greater than 0."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}